+++ frontend/bindings/case/internal/app/index.ts
@@ -7,6 +7,7 @@ export {
 };
 
 export type {
+    AttachmentDTO,
     PerformerDTO,
     PerformersPageDTO,
     SceneDTO,