+++ main.go
@@ -23,7 +23,7 @@ func main() {
 		},
 		Assets: application.AssetOptions{
 			Handler:    application.AssetFileServerFS(assets),
-			Middleware: app.CoverMiddleware(appService),
+			Middleware: app.AssetMiddleware(appService),
 		},
 		Mac: application.MacOptions{
 			ApplicationShouldTerminateAfterLastWindowClosed: true,